1.15.7 refactor: dont need duplicate event classes for player features
